+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.8.0.1
 
+- Long track and playlist names end in … and scroll when you hover over them.
 - Fixed a small shake with Direction of travel as the camera comes to a stop at a hold or the end of a track.
 
 ## 0.7.0.1